Check if skeleton is directory while creating home folder
Allow /dev/null as skeleton.

+++ lib/ansible/modules/user.py
@@ -1247,7 +1247,7 @@ class User(object):
             else:
                 skeleton = '/etc/skel'
 
-            if os.path.exists(skeleton):
+            if os.path.exists(skeleton) and os.path.isdir(skeleton):
                 try:
                     shutil.copytree(skeleton, path, symlinks=True)
                 except OSError as e:
